requirements.txt
Flask
gunicorn
requests
google-generativeai